requirements.txt

django-model-search
